+++ python/cugraph/tests/test_bfs.py
@@ -36,14 +36,6 @@ with warnings.catch_warnings():
 # =============================================================================
 DIRECTED_GRAPH_OPTIONS = [True, False]
 
-TINY_DATASETS = ['../datasets/karate.csv',
-                 '../datasets/dolphins.csv',
-                 '../datasets/polbooks.csv']
-SMALL_DATASETS = ['../datasets/netscience.csv',
-                  '../datasets/email-Eu-core.csv']
-
-DATASETS = TINY_DATASETS + SMALL_DATASETS
-
 SUBSET_SEED_OPTIONS = [42]
 
 DEFAULT_EPSILON = 1e-6
@@ -243,7 +235,7 @@ def _compare_bfs_spc(G, Gnx, source):
 # =============================================================================
 # Tests
 # =============================================================================
-@pytest.mark.parametrize('graph_file', DATASETS)
+@pytest.mark.parametrize('graph_file', utils.DATASETS_5)
 @pytest.mark.parametrize('directed', DIRECTED_GRAPH_OPTIONS)
 @pytest.mark.parametrize('seed', SUBSET_SEED_OPTIONS)
 def test_bfs(graph_file, directed, seed):
@@ -253,7 +245,7 @@ def test_bfs(graph_file, directed, seed):
                 seed=seed)
 
 
-@pytest.mark.parametrize('graph_file', DATASETS)
+@pytest.mark.parametrize('graph_file', utils.DATASETS)
 @pytest.mark.parametrize('directed', DIRECTED_GRAPH_OPTIONS)
 @pytest.mark.parametrize('seed', SUBSET_SEED_OPTIONS)
 def test_bfs_spc(graph_file, directed, seed):
@@ -263,7 +255,7 @@ def test_bfs_spc(graph_file, directed, seed):
                 seed=seed)
 
 
-@pytest.mark.parametrize('graph_file', TINY_DATASETS)
+@pytest.mark.parametrize('graph_file', utils.TINY_DATASETS)
 @pytest.mark.parametrize('directed', DIRECTED_GRAPH_OPTIONS)
 def test_bfs_spc_full(graph_file, directed):
     """Test BFS traversal on every vertex with shortest path counting"""